flask/requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
factory-boy==3.3.0
Faker==26.0.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
packaging==24.1
pluggy==1.5.0
PuLP==2.8.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
pytest==8.2.2
pytest-sqlalchemy-mock==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==70.1.1
six==1.16.0
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
Werkzeug==3.0.3
wheel==0.43.0
